vendorcode/intel/edk2: Fix EFI_PEI_GRAPHICS_INFO_HOB structure
Change the FrameBufferSize field from UINT64 to UINT32 to match the
Platform Initialization 2.4 specification.

BRANCH=none
BUG=None
TEST=Build and run on cyan

Reviewed-on: http://review.coreboot.org/10834
Tested-by: build bot (Jenkins)

+++ src/vendorcode/intel/edk2/uefi_2.4/uefi_types.h
@@ -78,10 +78,10 @@ static inline void debug_dead_loop(void)
 //
 // Contents of the PEI_GRAPHICS_INFO_HOB
 //
-typedef struct  {
-	EFI_PHYSICAL_ADDRESS FrameBufferBase;
-	UINT64 FrameBufferSize;
-	EFI_GRAPHICS_OUTPUT_MODE_INFORMATION GraphicsMode;
+typedef struct {
+	EFI_PHYSICAL_ADDRESS			FrameBufferBase;
+	UINT32					FrameBufferSize;
+	EFI_GRAPHICS_OUTPUT_MODE_INFORMATION	GraphicsMode;
 } EFI_PEI_GRAPHICS_INFO_HOB;
 
 //
